+++ nix/launcher-config.nix
@@ -2,25 +2,29 @@
 , environment ? "staging"
 , os ? "linux"
 , jormungandrLib ? (import ../. {}).jormungandrLib
+, cardanoLib
 , runCommand
 , lib
 , devShell ? false
 , cardano-wallet-native
+, runCommandNative
 }:
 let
   dirSep = if os == "windows" then "\\" else "/";
-  envCfg = jormungandrLib.environments.${environment};
+  envCfg = (if (backend == "cardano") then cardanoLib else jormungandrLib).environments.${environment};
   cfg = jormungandrLib.mkConfig (envCfg // {
     trustedPeers = envCfg.daedalusPeers or envCfg.trustedPeers;
   });
   jormungandrConfigForCluster = builtins.toFile "jormungandr-config-${environment}.yaml" (builtins.toJSON cfg);
 
+  installDirectorySuffix.testnet = "Testnet";
+  installDirectorySuffix.staging = "Staging";
   installDirectorySuffix.qa = "QA";
   installDirectorySuffix.selfnode = "SelfNode";
   installDirectorySuffix.nightly = "Nightly";
   installDirectorySuffix.itn_rewards_v1 = "- Rewards v1";
 
-  spacedName = "Daedalus ${installDirectorySuffix.${environment}}";
+  spacedName = if environment == "mainnet" then "Daedalus" else "Daedalus ${installDirectorySuffix.${environment}}";
 
   dataDir.linux = "\${XDG_DATA_HOME}/Daedalus/${environment}";
   dataDir.macos64 = "\${HOME}/Library/Application Support/${spacedName}";
@@ -63,7 +67,7 @@ let
   configPath.windows = "\${DAEDALUS_INSTALL_DIRECTORY}\\config.yaml";
   configPath.macos64 = "\${DAEDALUS_INSTALL_DIRECTORY}/../Resources/config.yaml";
 
-  selfnodeBlock0 = runCommand "selfnode-block0.bin" { buildInputs = [ cardano-wallet-native.jormungandr-cli ]; } ''
+  selfnodeBlock0 = runCommandNative "selfnode-block0.bin" { buildInputs = [ cardano-wallet-native.jormungandr-cli ]; } ''
     jcli genesis encode --input ${genesisPath.linux} --output $out
   '';
 
@@ -73,6 +77,28 @@ let
     linux = jormungandrConfigForCluster;
   };
   finalJormungandrCfgPath = if devShell then jormungandrConfigForCluster else cfgPathForOs.${os};
+
+  tier2-cfg-files = runCommand "tier2-cfg-files" {
+    nodeConfig = builtins.toJSON (envCfg.nodeConfig // { GenesisFile = finalGenesisLocation.${os}; });
+    topologyFile = cardanoLib.mkEdgeTopology {
+      inherit (envCfg) edgePort;
+      edgeNodes = [ envCfg.relaysNew ];
+    };
+    passAsFile = [ "nodeConfig" ];
+  } ''
+    mkdir $out
+    cp ${envCfg.genesisFile} $out/${environment}-genesis.json
+    cp $nodeConfigPath $out/configuration-${environment}.yaml
+    cp $topologyFile $out/${environment}-topology.yaml
+  '';
+
+  finalGenesisLocation.linux = envCfg.genesisFile;
+  finalGenesisLocation.macos64 = envCfg.genesisFile;
+
+  byronConfigDir = {
+    linux = tier2-cfg-files;
+    macos64 = tier2-cfg-files;
+  };
 
   walletArgs = [
     "launch" ] ++
@@ -100,12 +126,6 @@ let
   ];
 
   launcherConfig = {
-    walletBin = walletBin.${os};
-    walletArgs = if environment == "selfnode" then walletArgsSelfnode else walletArgs;
-
-    nodeBin = nodeBin.${os};
-    nodeArgs = [];
-
     daedalusBin = daedalusBin.${os};
     walletLogging = true;
     stateDir = dataDir.${os};
@@ -125,15 +145,47 @@ let
     updaterPath = "/foo";
     updaterArgs = [];
     updateArchive = "/bar";
-  };
+  } // (lib.optionalAttrs (backend == "cardano") {
+    networkName = environment;
+    nodeConfig = {
+      kind = "byron";
+      configurationDir = byronConfigDir.${os};
+      network = {
+        configFile = "configuration-${environment}.yaml";
+        genesisFile = "${environment}-genesis.json";
+        genesisHash = envCfg.genesisHash;
+        topologyFile = "${environment}-topology.yaml";
+      };
+    };
+    syncTolerance = "300s";
+  }) // (lib.optionalAttrs (backend == "jormungandr") {
+    block0Path = if (envCfg ? block0bin) then block0Bin.${os} else "";
+    block0Hash = jormungandrLib.environments.${environment}.genesisHash;
+    configPath = finalJormungandrCfgPath;
+    walletBin = walletBin.${os};
+    walletArgs = if environment == "selfnode" then walletArgsSelfnode else walletArgs;
+    nodeBin = nodeBin.${os};
+    nodeArgs = [];
+    syncTolerance = if (environment == "selfnode") then "600s" else jormungandrLib.environments.${environment}.syncTolerance;
+  }) // (lib.optionalAttrs (environment == "selfnode") {
+    block0Path = if ((os == "linux") || devShell) then selfnodeBlock0 else block0Bin.${os};
+    block0Hash = builtins.replaceStrings ["\n"] [""] (builtins.readFile (runCommandNative "selfnode-block0.hash" { buildInputs = [ cardano-wallet-native.jormungandr-cli ]; } ''
+      jcli genesis hash --input ${selfnodeBlock0} > $out
+    ''));
+    secretPath = if devShell then secretPath.linux else secretPath.${os};
+    configPath = if devShell then configPath.linux else configPath.${os};
+  });
+
   hasBlock0 = (environment != "selfnode") && envCfg ? block0bin;
+
   installerConfig = {
     installDirectory = if os == "linux" then "Daedalus/${environment}" else spacedName;
     inherit spacedName;
     macPackageName = "Daedalus${environment}";
     dataDir = dataDir.${os};
+  } // (lib.optionalAttrs (backend == "jormungandr") {
     inherit hasBlock0;
-  } // (lib.optionalAttrs hasBlock0 {
+  }) // (lib.optionalAttrs ((backend == "jormungandr") && hasBlock0) {
     block0 = envCfg.block0bin;
   }) // (lib.optionalAttrs (environment == "selfnode") {
     genesisPath = genesisPath.linux;
@@ -152,13 +204,17 @@ in {
   } ''
     mkdir $out
     cd $out
-    ${if (environment == "selfnode") then ''
-      cp ${../utils/jormungandr/selfnode/config.yaml} config.yaml
-      cp ${../utils/jormungandr/selfnode/genesis.yaml} genesis.yaml
-      cp ${../utils/jormungandr/selfnode/secret.yaml} secret.yaml
-    '' else "cp ${jormungandrConfigForCluster} jormungandr-config.yaml"}
+    ${if (backend == "jormungandr") then ''
+      ${lib.optionalString (installerConfig.hasBlock0) "cp ${installerConfig.block0} block-0.bin"}
+      ${if (environment == "selfnode") then ''
+        cp ${../utils/jormungandr/selfnode/config.yaml} config.yaml
+        cp ${../utils/jormungandr/selfnode/genesis.yaml} genesis.yaml
+        cp ${../utils/jormungandr/selfnode/secret.yaml} secret.yaml
+      '' else "cp ${jormungandrConfigForCluster} jormungandr-config.yaml"}
+    ''
+    else ""
+    }
     cp $installerConfigPath installer-config.json
     cp $launcherConfigPath launcher-config.yaml
-    ${lib.optionalString (installerConfig.hasBlock0) "cp ${installerConfig.block0} block-0.bin"}
   '';
 }